Move actions file into direct-funding-store
